backend/requirements.txt
fastapi
uvicorn
snowflake-connector-python[pandas]
snowflake-sqlalchemy
python-dotenv